requirements.txt
gql
requests-toolbelt
psycopg2
pandas
geopandas